@@ -1,16 +1,15 @@
 name: alice_lightweight
 description: Lightweight Alice version. Removed additional dependencies except dio.
-version: 3.4.1
+version: 3.5.1
 homepage: https://github.com/jhomlala/alice
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.13.5
-  dio: ^5.1.1
+  dio: ^5.3.3
   rxdart: ^0.27.7
-  share_plus: ^6.3.1
-  path_provider: ^2.0.14
+  share_plus: ^7.1.0
+  path_provider: ^2.1.1
