@@ -115,7 +115,7 @@ class AliceDioInterceptor extends InterceptorsWrapper {
   /// Handles error and adds data to alice http call
   @override
   void onError(
-    DioError error,
+    DioException error,
     ErrorInterceptorHandler handler,
   ) {
     var httpError = AliceHttpError();
