@@ -1,8 +1,6 @@
 import 'dart:io';
-import 'package:alice_lightweight/core/alice_http_adapter.dart';
 import 'package:alice_lightweight/model/alice_http_call.dart';
 
-import 'package:http/http.dart' as http;
 import 'package:alice_lightweight/core/alice_core.dart';
 import 'package:alice_lightweight/core/alice_dio_interceptor.dart';
 import 'package:alice_lightweight/core/alice_http_client_adapter.dart';
@@ -16,10 +14,13 @@ class Alice {
   GlobalKey<NavigatorState> _navigatorKey;
   AliceCore _aliceCore;
   AliceHttpClientAdapter _httpClientAdapter;
-  AliceHttpAdapter _httpAdapter;
 
-  Alice._(this.darkTheme, this._navigatorKey, this._aliceCore,
-      this._httpClientAdapter, this._httpAdapter);
+  Alice._(
+    this.darkTheme,
+    this._navigatorKey,
+    this._aliceCore,
+    this._httpClientAdapter,
+  );
 
   /// Creates alice instance.
   factory Alice({
@@ -29,14 +30,12 @@ class Alice {
   }) {
     final aliceCore = AliceCore(navigatorKey, darkTheme);
     final httpClientAdapter = AliceHttpClientAdapter(aliceCore);
-    final httpAdapter = AliceHttpAdapter(aliceCore);
 
     return Alice._(
       darkTheme,
       navigatorKey,
       aliceCore,
       httpClientAdapter,
-      httpAdapter,
     );
   }
 
@@ -65,11 +64,6 @@ class Alice {
       HttpClientResponse response, HttpClientRequest request,
       {dynamic body}) {
     _httpClientAdapter.onResponse(response, request, body: body);
-  }
-
-  /// Handle both request and response from http package
-  void onHttpResponse(http.Response response, {dynamic body}) {
-    _httpAdapter.onResponse(response, body: body);
   }
 
   /// Opens Http calls inspector. This will navigate user to the new fullscreen
